🛠️ refactor: Refactor App.jsx and redux files.
